+++ src/material/mod.rs
@@ -1,12 +1,12 @@
 //! Implements material models
 
+mod conductivity;
 mod linear_elastic;
 mod local_state;
-mod model_conductivity;
-mod model_stress_strain;
+mod stress_strain;
 mod von_mises;
+pub use crate::material::conductivity::*;
 pub use crate::material::linear_elastic::*;
 pub use crate::material::local_state::*;
-pub use crate::material::model_conductivity::*;
-pub use crate::material::model_stress_strain::*;
+pub use crate::material::stress_strain::*;
 pub use crate::material::von_mises::*;
